requirements.txt
groq
deepgram-sdk
sounddevice
python-dotenv
customtkinter
numpy
soundfile
edge-tts
elevenlabs
